requirements.txt
Django==5.2.3
djangorestframework==3.16.0
Markdown==3.8
django-filter==25.1